% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refset.R
\name{is.parcel}
\alias{is.parcel}
\title{Checks whether an object is a parcel}
\usage{
is.parcel(x)
}
\arguments{
\item{x}{an object to examine}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks whether an object is a parcel
}
\seealso{
Other wrapping functions: \code{\link{contents}},
  \code{\link{unwrap_as}}, \code{\link{wrapset}},
  \code{\link{wrap}}
}
