% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refset.R
\docType{package}
\name{refset-package}
\alias{refset-package}
\title{Subsets with Reference Semantics}
\description{
Subsets with Reference Semantics
}
\details{
Refset provides subsets with reference semantics, i.e. subsets
which automatically reflect changes in the original object, and which
also update the original object when they are changed.

For more info see \code{vignette('refset')}.
}
\author{
David Hugh-Jones \email{davidhughjones@gmail.com}
}
\keyword{package}
