% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refset.R
\name{unwrap_as}
\alias{unwrap_as}
\title{Unwrap contents of a parcel into a new variable}
\usage{
unwrap_as(x, parcel, env = parent.frame())
}
\arguments{
\item{x}{name of the variable to bind to}

\item{parcel}{an object of class 'parcel'}

\item{env}{environment to assign the variable into}
}
\description{
\code{unwrap_as} creates a new variable which, when evaluated,
calls \code{\link{contents}} to return the parcel contents.
}
\examples{
vec <- 1:10
parcel <- wrapset(vec, vec > 3) 
unwrap_as(y, parcel)
y
}
\seealso{
Other wrapping functions: \code{\link{contents}},
  \code{\link{is.parcel}}, \code{\link{wrapset}},
  \code{\link{wrap}}
}
