% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wrap}
\alias{wrap}
\title{Wrap an expression and its environment into a parcel.}
\usage{
wrap(expr, env = parent.frame())
}
\arguments{
\item{expr}{an R expression}

\item{env}{environment in which \code{expr} is to be evaluated}
}
\value{
An object of class 'parcel', with components \code{expr} and \code{env}.
}
\description{
Refsets (and other active bindings) cannot be passed as function
arguments, since doing so makes a copy. \code{wrap} allows you to pass
arbitrary expressions between functions.
}
\examples{
dfr <- data.frame(a=1:5, b=1:5)
rs \%r\% dfr[1:2,]
parcel <- wrap(rs)
f <- function (parcel) contents(parcel) <- contents(parcel)*2
f(parcel)
f(parcel)
}
\seealso{
To evaluate the parcel contents in the environment \code{env}, call
\code{\link{contents}} on the parcel. To change parcel contents use
\code{\link{contents<-}}.To rebind the contents to a variable, use
\code{\link{unwrap_as}}.
}

