% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/refset.R
\name{wrapset}
\alias{wrapset}
\title{Convenience function to create a parcel containing a refset.}
\usage{
wrapset(data, ..., env = parent.frame())
}
\arguments{
\item{data,...}{passed to \code{\link{refset}}}

\item{env}{passed to \code{\link{refset}} as argument \code{eval.env}}
}
\value{
A \code{parcel} object.
}
\description{
\code{wrapset} calls \code{\link{refset}} on its arguments and
returns the resulting active binding in a \code{parcel} object
for passing around.
}
\examples{
dfr <- data.frame(a=1:5, b=1:5)
parcel <- wrapset(dfr, a<3, , drop=FALSE)
contents(parcel)
}
\seealso{
Other wrapping functions: \code{\link{contents}},
  \code{\link{contents<-}}; \code{\link{is.parcel}};
  \code{\link{unwrap_as}}; \code{\link{wrap}}
}

